requirements.txt
aiosmtpd==1.1
telepot==12.7
pyyaml==3.12